Move config tests to separate file
